move no main funcs to internal
