+++ cmd/api/internal/otel.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package main
+package internal
 
 import (
 	"go.opentelemetry.io/otel"
@@ -29,7 +29,7 @@ import (
 	gcppropagator "github.com/GoogleCloudPlatform/opentelemetry-operations-go/propagator"
 )
 
-func newTracer(projectId string) (*sdktrace.TracerProvider, error) {
+func NewTracer(projectId string) (*sdktrace.TracerProvider, error) {
 
 	exporter, err := texporter.New(texporter.WithProjectID(projectId))
 	if err != nil {
